@@ -1,7 +1,7 @@
 {
   "devDependencies": {
     "@types/node": "^20.11.30",
-    "typescript": "^5.5.2",
+    "typescript": "^5.9.2",
     "xlsx": "^0.18.5"
   }
 }
