@@ -5,6 +5,28 @@
 
 namespace jsb
 {
+    struct TransferData
+    {
+        NativeObjectID source_worker_id;
+        uint32_t transfer_index;
+        Variant variant;
+        String script_path;
+        List<Pair<StringName, Variant>> state;
+
+        TransferData() : transfer_index(0)
+        {
+        }
+
+        TransferData(NativeObjectID p_source_worker_id, uint32_t p_transfer_index, const Variant& p_variant,
+                     const String& p_script_path = {}, const List<Pair<StringName, Variant>>& p_state = {})
+            : source_worker_id(p_source_worker_id), transfer_index(p_transfer_index), variant(p_variant),
+              script_path(p_script_path), state(p_state)
+        {
+        }
+
+        ~TransferData() = default;
+    };
+
     struct Message
     {
     public:
@@ -22,7 +44,7 @@ namespace jsb
             TYPE_ERROR,
         };
 
-        Message() = default;
+        Message() = delete;
         ~Message() = default;
 
         Message(const Message&) = delete;
@@ -31,8 +53,11 @@ namespace jsb
         Message(Message&&) noexcept = default;
         Message& operator=(Message&&) noexcept = default;
 
-        Message(Type p_type, NativeObjectID p_id, Buffer&& p_buffer)
-        : type_(p_type), id_(p_id), buffer_(std::move(p_buffer)) {}
+        Message(Type p_type, NativeObjectID p_id, Buffer&& p_buffer = Buffer(),
+            std::vector<TransferData>&& p_transfers = std::vector<TransferData>())
+            : type_(p_type), id_(p_id), buffer_(std::move(p_buffer)), transfers(std::move(p_transfers))
+        {
+        }
 
         // object id of worker object in master env
         NativeObjectID get_id() const { return id_; }
@@ -41,10 +66,13 @@ namespace jsb
 
         const Buffer& get_buffer() const { return buffer_; }
 
+        const std::vector<TransferData>& get_transfers() const { return transfers; }
+
     private:
         Type type_;
         NativeObjectID id_;
         Buffer buffer_;
+        std::vector<TransferData> transfers;
     };
 
 }
