@@ -95,12 +95,14 @@ namespace jsb::internal
 
     String PathUtil::convert_javascript_path(const String& p_source_path)
     {
-        if (p_source_path.ends_with("." JSB_JAVASCRIPT_EXT))
+        bool is_js_extension = p_source_path.ends_with("." JSB_JAVASCRIPT_EXT);
+        if (is_js_extension || p_source_path.ends_with("." JSB_COMMONJS_EXT))
         {
+            int extension_length = is_js_extension ? 3 : 4;
             const String root_path = Settings::get_jsb_out_res_path();
             jsb_checkf(p_source_path.begins_with(root_path), "can not proceed javascript sources not under the project data directory");
             const String replaced = String("res://").path_join(
-                p_source_path.substr(root_path.length() + 1, p_source_path.length() - root_path.length() - 3)
+                p_source_path.substr(root_path.length() + 1, p_source_path.length() - root_path.length() - extension_length)
                 + JSB_TYPESCRIPT_EXT);
             return replaced;
         }
