@@ -11,30 +11,27 @@ function pascal_to_upper_snake_case(str: string) {
     return str.replace(/[a-z][A-Z]|[0-9][A-Z][a-z]/g, (m) => `${m[0]}_${m.slice(1)}`).toUpperCase();
 }
 
-function bind(target: any, value: any) {
-    return typeof value === "function" ? value.bind(target) : value;
-}
-
 function is_basic_object(value: object) {
     const proto = Object.getPrototypeOf(value);
     return proto === Object.prototype || !proto;
 }
 
-function proxy_value(value: any) {
+function proxy_unwrap_value<T>(value: T): T {
+    return (value as any)?.[ProxyTarget] ?? value;
+}
+
+function proxy_wrap_value<T>(value: T): T;
+function proxy_wrap_value(value: any): any {
     if (value == null) {
         return value;
     }
 
-    if (value[ProxyTarget]) {
+    if ((value as any)[ProxyTarget]) {
         return value;
     }
 
     if (typeof value === "function") {
-        const proxied_function = function(this: any, ...args: any[]) {
-            return proxy_value(value.apply(this, args.map(proxy_value)));
-        } as any as (Function & { [ProxyTarget]: Function });
-        proxied_function[ProxyTarget] = value;
-        return proxied_function;
+        return function_proxy(value);
     }
 
     if (typeof value !== "object") {
@@ -55,31 +52,94 @@ const object_handler = {
         }
 
         const value = Reflect.get(target, p);
+        const descriptor = Object.getOwnPropertyDescriptor(target, p);
 
-        if (typeof p !== "string") {
+        if (typeof p !== "string" || (descriptor && !descriptor.writable && !descriptor.configurable)) {
             return value;
         }
 
         if (p[0]?.toUpperCase() === p[0]
-          && value
-          && typeof value === "object"
-          && !Array.isArray(target)
-          && is_basic_object(value)
-          && !Object.entries(value).find(([k, v]) => k[0].toUpperCase() !== k[0] || typeof v !== "number")) {
+            && value
+            && typeof value === "object"
+            && is_basic_object(value)
+            && !Object.entries(value).find(([k, v]) => k[0].toUpperCase() !== k[0] || typeof v !== "number")) {
             return enum_proxy(value);
         }
 
-        return proxy_value(bind(target, value));
+        return proxy_wrap_value(value);
+    },
+    getOwnPropertyDescriptor(target, p) {
+        return Reflect.getOwnPropertyDescriptor(target, typeof p === "string" ? get_member(p) : p)
+            ?? Reflect.getOwnPropertyDescriptor(target, p);
+    },
+    has(target, p) {
+        return Reflect.has(target, typeof p === "string" ? get_member(p) : p)
+            || Reflect.has(target, p);
+    },
+    ownKeys(target) {
+        return Reflect.ownKeys(target).map(key => typeof key === "string" && get_internal_mapping(key) || key);
+    },
+    set(target, p, new_value, _receiver) {
+        return Reflect.set(target, typeof p === "string" ? get_member(p) : p, new_value);
     },
 } satisfies ProxyHandler<any>;
 
+const object_properties_handler = {
+    ...object_handler,
+    get(target, p, _receiver) {
+        if (p === ProxyTarget) {
+            return target;
+        }
 
-function array_proxy<T>(arr: T): T {
+        const descriptor = Object.getOwnPropertyDescriptor(target, p);
+
+        if (typeof p !== "string" || (descriptor && !descriptor.writable && !descriptor.configurable)) {
+            return Reflect.get(target, p);
+        }
+
+        const value = Reflect.get(target, p !== "toString" ? get_member(p) : p);
+
+        if (p[0]?.toUpperCase() === p[0]
+            && value
+            && typeof value === "object"
+            && is_basic_object(value)
+            && !Object.entries(value).find(([k, v]) => k[0].toUpperCase() !== k[0] || typeof v !== "number")) {
+            return enum_proxy(value);
+        }
+
+        return proxy_wrap_value(value);
+    },
+} satisfies ProxyHandler<any>;
+
+function array_proxy<T extends any[]>(arr: T): T {
     return new Proxy(arr, object_handler);
 }
 
-function object_proxy<T>(obj: T): T {
-    return new Proxy(obj, object_handler);
+function object_proxy<T extends object>(obj: T, remap_properties?: boolean): T {
+    return new Proxy(obj, remap_properties ? object_properties_handler : object_handler);
+}
+
+const key_only_handler = {
+    apply: function(target, this_arg, args) {
+        return target.apply(this_arg?.[ProxyTarget] ?? this_arg, args);
+    },
+    get(target, p, _receiver) {
+        if (p === ProxyTarget) {
+            return target;
+        }
+
+        const descriptor = Object.getOwnPropertyDescriptor(target, p);
+
+        if (typeof p !== "string" || (descriptor && !descriptor.writable && !descriptor.configurable)) {
+            return Reflect.get(target, p);
+        }
+
+        return key_only_proxy(Reflect.get(target, p !== "toString" ? get_member(p) : p));
+    },
+} satisfies ProxyHandler<any>;
+
+function key_only_proxy<T extends object | ((...args: any[]) => any)>(target: T) {
+    return new Proxy(target, key_only_handler);
 }
 
 const instance_handler = {
@@ -98,7 +158,7 @@ const instance_handler = {
             return Reflect.get(target, p);
         }
 
-        return proxy_value(bind(target, Reflect.get(target, p !== "toString" ? get_member(p) : p)));
+        return proxy_wrap_value(Reflect.get(target, p !== "toString" ? get_member(p) : p));
     },
     getOwnPropertyDescriptor(target, p) {
         return Reflect.getOwnPropertyDescriptor(target, typeof p === "string" ? get_member(p) : p);
@@ -123,7 +183,7 @@ const instance_handler = {
     },
 } satisfies ProxyHandler<any>;
 
-function instance_proxy<T>(target_instance: T): T {
+function instance_proxy<T extends object>(target_instance: T): T {
     return new Proxy(target_instance, instance_handler);
 }
 
@@ -145,25 +205,42 @@ const class_handler = {
             return Reflect.get(target, p);
         }
 
+        const descriptor = Object.getOwnPropertyDescriptor(target, p);
+
+        if (descriptor && !descriptor.writable && !descriptor.configurable) {
+            return Reflect.get(target, p);
+        }
+
         if (p === "prototype") {
             const proto = Reflect.get(target, "prototype");
             return proto && class_proxy(proto);
         }
 
         if (p[0]?.toUpperCase() !== p[0]) {
-            return proxy_value(bind(target, Reflect.get(target, get_member(p))));
+            return proxy_wrap_value(Reflect.get(target, get_member(p)));
         }
 
         if (p.toUpperCase() === p) {
-            return proxy_value(bind(target, Reflect.get(target, p)));
+            return proxy_wrap_value(Reflect.get(target, p));
         }
 
         return enum_proxy(Reflect.get(target, get_enum(p)));
     },
 } satisfies ProxyHandler<any>;
 
-function class_proxy<T>(target_class: T): T {
+function class_proxy<T extends object>(target_class: T): T {
     return new Proxy(target_class, class_handler);
+}
+
+const function_handler = {
+    ...class_handler,
+    apply: function(target, this_arg, args) {
+        return proxy_wrap_value(target.apply(this_arg?.[ProxyTarget] ?? this_arg, args.map(proxy_wrap_value)));
+    },
+} satisfies ProxyHandler<any>;
+
+function function_proxy<T extends (...args: any[]) => any>(fn: T): T {
+    return new Proxy(fn, function_handler);
 }
 
 const enum_handler = {
@@ -182,7 +259,7 @@ const enum_handler = {
             return Reflect.get(target, p);
         }
 
-        return bind(target, Reflect.get(target, get_enum_value(p)));
+        return Reflect.get(target, get_enum_value(p));
     },
     getOwnPropertyDescriptor(target, p) {
         return Reflect.getOwnPropertyDescriptor(target, typeof p === "string" ? get_enum_value(p) : p);
@@ -207,7 +284,7 @@ const enum_handler = {
     },
 } satisfies ProxyHandler<any>;
 
-function enum_proxy<T>(target_enum: T): T {
+function enum_proxy<T extends object>(target_enum: T): T {
     if (typeof target_enum !== "object") {
         return target_enum;
     }
@@ -236,7 +313,7 @@ const api_handler = (target: any) => ({
         }
 
         if (p === "toString") {
-            return proxy_value(bind(target, Reflect.get(target, p)));
+            return proxy_wrap_value(Reflect.get(target, p));
         }
 
         // Special case, see jsb_godot_module_loader.cpp
@@ -245,7 +322,7 @@ const api_handler = (target: any) => ({
         }
 
         if (p[0]?.toUpperCase() !== p[0]) {
-            return proxy_value(bind(target, Reflect.get(target, get_member(p))));
+            return proxy_wrap_value(Reflect.get(target, get_member(p)));
         }
 
         const value = Reflect.get(target, get_class(p));
@@ -289,8 +366,28 @@ const api_handler = (target: any) => ({
     },
 } satisfies ProxyHandler<any>);
 
-const godot_jsb = object_proxy(jsb_api) as typeof GodotJsb;
-const api = new Proxy({ jsb: godot_jsb }, api_handler(godot_api)) as typeof Godot & { jsb: typeof GodotJsb };
+const proxy = {
+    array_proxy,
+    class_proxy,
+    enum_proxy,
+    function_proxy,
+    instance_proxy,
+    key_only_proxy,
+    object_proxy,
+    proxy_unwrap_value,
+    proxy_wrap_value,
+};
+
+type GodotLibApi = typeof Godot & {
+    jsb: typeof GodotJsb;
+    proxy: typeof proxy;
+}
+
+const jsb = object_proxy(jsb_api) as typeof GodotJsb;
+const api = new Proxy({
+    jsb,
+    proxy,
+}, api_handler(godot_api)) as GodotLibApi;
 
 /**
  * This is a starting point for writing GodotJS code that is camel-case binding agnostic at runtime.
