@@ -1,11 +1,10 @@
-
 declare module "godot.worker" {
-    import { Object as GDObject } from "godot";
+    import { GAny, GArray, Object as GObject } from "godot";
 
     class JSWorker {
         constructor(path: string);
 
-        postMessage(message: any): void;
+        postMessage(message: any, transfer?: GArray | ReadonlyArray<NonNullable<GAny>>): void;
         terminate(): void;
 
         onready?: () => void;
@@ -14,7 +13,11 @@ declare module "godot.worker" {
         //TODO not implemented yet
         onerror?: (error: any) => void;
 
-        ontransfer?: (obj: GDObject) => void;
+        /**
+         * @deprecated Use onmessage to receive messages sent from postMessage() with transfers included.
+         * @param obj
+         */
+        ontransfer?: (obj: GObject) => void;
     }
 
     // only available in worker scripts
@@ -23,9 +26,13 @@ declare module "godot.worker" {
         
         close(): void,
 
-        transfer(obj: GDObject): void,
+        /**
+         * @deprecated Use the transfer parameter of postMessage instead.
+         * @param obj
+         */
+        transfer(obj: GObject): void,
 
-        postMessage(message: any): void,
+        postMessage(message: any, transfer?: GArray | ReadonlyArray<NonNullable<GAny>>): void;
 
     } | undefined;
 
