@@ -53,8 +53,8 @@ private:
 
     std::shared_ptr<jsb::internal::Process> tsc_;
 
-    void _generate_edited_scene_dts(const String& p_path);
-    void _generate_edited_resource_dts(const Ref<Resource>& p_resource);
+    void _on_scene_saved(const String& p_path);
+    void _on_resource_saved(const Ref<Resource>& p_resource);
     void _generate_imported_resource_dts(const Vector<String>& p_resources);
 
 protected:
@@ -75,8 +75,8 @@ protected:
     static bool delete_file(const String& p_file);
     static void get_all_scenes(EditorFileSystemDirectory* p_dir, Vector<String>& r_list);
     static void get_all_resources(EditorFileSystemDirectory* p_dir, Vector<String>& r_list);
-    static void generate_scene_nodes_dts(const Vector<String>& p_paths);
-    static void generate_resource_dts(const Vector<String>& p_paths);
+    static void generate_scene_nodes_types(const Vector<String>& p_paths);
+    static void generate_resource_types(const Vector<String>& p_paths);
 
 public:
     GodotJSEditorPlugin();
@@ -91,14 +91,13 @@ public:
     bool verify_ts_project() const;
     void _ignore_node_modules();
     void cleanup_invalid_files();
-    void generate_edited_scene_dts();
 
     // not really a singleton, but always get from `EditorNode` which assumed unique
     static GodotJSEditorPlugin* get_singleton();
 
     static void generate_godot_dts();
-    static void generate_all_scene_nodes_dts();
-    static void generate_all_resource_dts();
+    static void generate_all_scene_nodes_types();
+    static void generate_all_resource_types();
     static void ignore_node_modules();
     static void collect_invalid_files(Vector<String>& r_invalid_files);
     static void collect_invalid_files(const String& p_path, Vector<String>& r_invalid_files);
