@@ -30,11 +30,11 @@ protected:
 private:
     static HashSet<String> ignored_paths_;
 
-    bool export_compiled_script(const String& p_path);
-    bool export_module_files(const jsb::JavaScriptModule& p_module);
-    bool export_raw_file(const String& p_path);
+    bool export_compiled_script(const String& p_path, bool p_remap);
+    bool export_module_files(const jsb::JavaScriptModule& p_module, bool p_remap);
+    bool export_raw_file(const String& p_path, bool p_remap);
     void export_raw_files(const PackedStringArray& p_paths, bool p_permit_typescript);
-    void get_script_resources(const String &p_dir, Vector<String> &r_list);
+    void get_script_resources(const String &p_dir, Vector<String> &r_list, bool p_is_node_module = false);
 
     HashSet<String> exported_paths_;
     std::shared_ptr<jsb::Environment> env_;
